feat: Add map links to site configuration
Introduces map link fields for both Solo and Blora office locations within the site configuration. This enhancement allows for direct linking to the respective office locations on mapping services, improving user navigation and accessibility.

Additionally, implements a global scroll-to-top functionality that triggers on route changes, ensuring a consistent and user-friendly browsing experience by resetting the scroll position to the top of the page for every new view.